backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==44.0.1

pydantic==2.5.0

python-multipart==0.0.6

python-dotenv==1.0.0

loguru==0.7.2

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2




